requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-pagination==0.12.17
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
